Move tests into packages
