+++ src/test/java/org/opentripplanner/updater/trip/moduletests/delay/SkippedTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater.trip.moduletests;
+package org.opentripplanner.updater.trip.moduletests.delay;
 
 import static com.google.transit.realtime.GtfsRealtime.TripDescriptor.ScheduleRelationship.SCHEDULED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -17,7 +17,7 @@ import org.opentripplanner.transit.model.timetable.TripTimes;
 import org.opentripplanner.updater.trip.GtfsRealtimeTestEnvironment;
 import org.opentripplanner.updater.trip.TripUpdateBuilder;
 
-public class B02_SkippedTest {
+public class SkippedTest {
 
   @Test
   public void scheduledTripWithSkippedAndScheduled() {
